requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
cloudinary==1.30.0
dj-database-url==1.0.0
Django==4.1.3
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-extensions==3.2.1
django-on-heroku==1.1.2
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
llvmlite==0.38.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.13
whitenoise==6.2.0